requirements.txt
```
streamlit
pypdf
anthropic
python-dotenv
